requirements.txt
bitstring==3.1.5
bitarray==0.8.3
pygame==1.9.4
Pillow==5.3.0
PySimpleGUI==3.20.0
